feat: a link preview built for the shape platforms crop to, and bulk vote orders
## The 1200x630 OG card

`FlierService::ogCard()`, served at …/{slug}/card.png, is now the og:image on both
the ballot and the flier page.

The flier is 4:5. Facebook and LinkedIn crop an og:image to 1.91:1 and WhatsApp
to roughly square, so the flier's bottom third — the vote URL, the rally copy and
the jury footnote — was cut off in every preview, on the platform's most-shared
surface. The card is a horizontal split instead: the face in a 480px column
(Cloudinary's new face-anchored `og_photo` preset), everything else beside it.
Nothing is cropped away because the ratio is already theirs.

Sized for ~380px, which is about how wide a preview renders in a WhatsApp thread.
Three elements survive that — name, gold rank chip, standing line — and the rest
is deliberately secondary. Momentum is omitted: a fourth number at 8 effective
pixels costs the other three their clarity. The flier is unchanged and is still
what a nominee downloads.

Three geometry bugs found by rendering it and looking, each producing a perfectly
valid PNG:

  • the name collided with the kicker, because the size ladder was copied from
    the flier's 952px column into a 608px one. Sizes are measured now (fitLines).
  • a surname wider than the column ran off the edge, because wrapMeasured places
    an over-wide word rather than break it and its output was trusted without
    checking widths. "Wolde-Giorgis" clears 608px at 76pt.
  • a name with stacked Yoruba diacritics grew UPWARD toward the kicker, because
    the baseline came from the point size, not the glyphs' ink extent. Measured:
    84px vs 79px for Ọlásùnkànmí against Olasunkanmi at 76pt.

The last two existed latently in the flier's own renderer; both are fixed there
too, as was the flier's programme line running underneath its rank chip.

The photo/panel seam fades per pixel rather than per column, because the
background is a vertical gradient — one flat colour leaves a bright band at one
end, which on a link preview is the entire first impression. Asserted as a
<=6-level step across the seam at three heights.

## "Can the site take over 1000 votes at a time?"

It could not, and it failed silently in three places.

`MAX_QTY = 1000` was applied as `min(MAX_QTY, $qty)`, so an order for 5,000
became an order for 1,000 — charged for 1,000, delivered 1,000, told nothing. The
buyer would learn it from the tally.

**gates_votes.weight was SMALLINT UNSIGNED.** A paid order mints ONE row with
weight = quantity, so 65,535 was the real ceiling, and it was invisible only
because MAX_QTY kept anything from reaching it. Raising the cap without fixing
this gives one of two outcomes: under strict mode the mint aborts (customer paid,
no votes), and on a host that relaxes sql_mode MySQL CLAMPS and reports success —
65,535 votes credited for an order of 100,000, no error anywhere. Widened to INT
UNSIGNED, matching vote_count and bonus_votes, which were already INT. That
migration is what actually makes bulk orders possible.

**amount_naira is INT UNSIGNED, and the ceilings multiply.** Quantity x an
admin-set rate can exceed it long before a quantity cap binds: at ₦1,000/vote a
10m-vote cap computes ₦10bn. So maxQtyForOrder() is the single authority — the
lower of the admin cap and what fits MAX_ORDER_NAIRA — read by the form's `max`,
the checkout's rejection, and the settings page's displayed figure, because three
surfaces computing it separately is how a form offers what the checkout refuses.

Now: `vote_max_qty` is an admin setting; an over-large order is REFUSED with the
real maximum rather than shrunk; and mint() carries its own storage guard, since
it is reachable from the gateway webhook long after checkout and the cap may have
moved. A refusal there leaves votes_used = 0 — the existing queryable "paid but
never minted, refund owed" signal, no new state to learn.

Worth stating for the client: an order of ANY size is one vote row and one
increment, so 50,000 votes is the same database work as one, and the binding
constraint on a genuinely large order is the payment gateway's per-transaction
limit, not this platform. Paid votes still move the public tally only —
organic_vote_count is asserted to stay at zero after a 50,000-vote mint.

docs/PAID-VOTE-CAPACITY.md covers the whole picture including the limits that
are not ours.

28 new tests; 1085 pass.

+++ database/schema.sql
@@ -126,7 +126,12 @@ CREATE TABLE IF NOT EXISTS gates_votes (
   voter_name VARCHAR(120) DEFAULT NULL,
   voter_phone VARCHAR(40) DEFAULT NULL,
   vote_type ENUM('standard','bonus','paid') NOT NULL DEFAULT 'standard',
-  weight SMALLINT UNSIGNED NOT NULL DEFAULT 1,
+  -- INT, not SMALLINT. A paid-vote order mints ONE row with weight = quantity, so
+  -- SMALLINT's 65,535 was the real (and unmeasured) ceiling on a bulk purchase — and
+  -- on a host that overrides sql_mode away from strict, MySQL would have CLAMPED to it
+  -- and reported success, crediting 65,535 votes for an order of 100,000. Matches
+  -- gates_nominees.vote_count and gates_donations.bonus_votes, which were already INT.
+  weight INT UNSIGNED NOT NULL DEFAULT 1,
   donation_id BIGINT UNSIGNED DEFAULT NULL,
   risk_score TINYINT UNSIGNED NOT NULL DEFAULT 0,
   fraud_flag TINYINT(1) NOT NULL DEFAULT 0,
